+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Changed
+
+- Replace `--log-map` with `--log-litterals`.
+
 ## [0.3.0]
 
 ### Added